requirements.txt
asgiref==3.5.0
cffi==1.15.0
cryptography==37.0.2
dj-database-url==0.5.0
Django==3.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pillow==9.1.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
